Add reliable mechanism for determining whether we are running under Cordova.
We were looking at window.Cordova, but that's always true when the
cordova.js file is loaded.

We were also looking at window.device, but that's only set after deviceready
is fired -- so we have race conditions.  That was causing a number of things
to break randomly, depending on whether we happened to win the race or not.

Adding a simple property from the Java side makes this 100% reliable.  I don't
know why Cordova upstream doesn't do this.