Recognize button click on touchstart for faster response time.
Android issues a 'touchcancel' event if the DOM is changed in the touchstart
processing.  Ways the DOM might be changed: audio playing (creates/clones
an element), Funf logging (XHR processing creates a <script> element maybe?).
So the hover effect doesn't persist for 'fast' buttons... but it's a reasonable
tradeoff here.
As a side-effect, we get a 'touchcancel' event